+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Support passing a block to `#clowne` for inline configuration. ([@palkan][])
+
 ## 0.1.0.beta1 (2018-01-08)
 
 - Initial version. ([@ssnickolay][], [@palkan][])